+++ popup/popup.js
@@ -1,13 +1,14 @@
 // On load of popup
 document.addEventListener("DOMContentLoaded", function () {
-  chrome.storage.sync.get(
-    ["AccessKeyId"],
-    function (data) {
-      if (typeof data.AccessKeyId !== "undefined")
-        document.getElementById("AccessKeyId").innerText = data.AccessKeyId;
-      chrome.storage.sync.clear();
-    }
-  );
+  chrome.storage.sync.get(["credentialsFile"], function (data) {
+    if (typeof data.credentialsFile !== "undefined")
+      document.getElementById("credentialsFile").innerText =
+        data.credentialsFile;
+  });
+  chrome.storage.sync.get(["env_variables"], function (data) {
+    if (typeof data.env_variables !== "undefined")
+      document.getElementById("env_variables").innerText =
+        data.env_variables;
+  });
+  chrome.storage.sync.clear();
 });
-
-